+++ attributes/poise-python.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-# https://github.com/poise/poise-python/issues/133
-default['poise-python']['options']['pip_version'] = '18.0'