+++ metadata.rb
@@ -5,14 +5,13 @@ maintainer       'Matt Kulka'
 maintainer_email 'matt@lqx.net'
 license          'MIT'
 description      'Procures Let\'s Encrypt SSL certificates for Route 53-hosted domains'
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.1.4'
+version          '2.0.0'
 
-supports     'ubuntu' if respond_to?(:supports)
+supports     'ubuntu'
 chef_version '>= 12'
 
-issues_url 'https://github.com/mattlqx/cookbook-letsencryptaws/issues' if respond_to?(:issues_url)
-source_url 'https://github.com/mattlqx/cookbook-letsencryptaws' if respond_to?(:source_url)
+issues_url 'https://github.com/mattlqx/cookbook-letsencryptaws/issues'
+source_url 'https://github.com/mattlqx/cookbook-letsencryptaws'
 
-depends 'poise-python', '~> 1.7'
-depends 'remote_file_s3', '~> 1.0.5'
+depends 'pyenv', '~> 3.2'
+depends 'remote_file_s3', '~> 1.0'