+++ spec/spec_helper.rb
@@ -55,15 +55,4 @@ RSpec.configure do |config|
 
   # Specify the Chef log_level (default: :warn)
   config.log_level = :warn
-
-  # Specify the path to a local JSON file with Ohai data
-  # config.path = 'ohai.json'
-
-  # Specify the operating platform to mock Ohai data from
-  config.platform = 'ubuntu'
-
-  # Specify the operating version to mock Ohai data from
-  config.version = '16.04'
 end
-
-at_exit { ChefSpec::Coverage.report! }